backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Background tasks
redis==5.0.1
celery==5.3.4
flower==2.0.1

# File storage (local filesystem used; no cloud clients required)

# Machine Learning and Data Science
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.3
biopython==1.81
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
joblib==1.3.2
shap==0.44.0

# Bioinformatics specific
Bio==1.6.0
pyfaidx==0.7.2.1
pyvcf3==1.0.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2
Pillow==10.1.0
reportlab==4.0.7
PyPDF2==3.0.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
